refac: Replace from 'Block' to 'byte'
- Block don't work functionally. It do like type data
- Clear nullable block data and air

+++ Assets/Scripts/BlockData.cs
@@ -8,7 +8,7 @@ namespace MyCraft
 {
     public struct BlockData
     {
-        public int id;
+        public byte id;
         public string name;
         public MaterialType materialType;
         public float hardness;
